# Backend/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EpiAirConsole API",
    "version": "1.0.0",
    "description": "API backend pour EpiAirConsole - gestion des utilisateurs et authentification"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Serveur de développement"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID unique de l'utilisateur"
          },
          "username": {
            "type": "string",
            "description": "Nom d'utilisateur"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email unique"
          },
          "avatar": {
            "type": "string",
            "description": "URL de l'avatar"
          },
          "gamesData": {
            "type": "array",
            "description": "Données de jeux de l'utilisateur"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de mise à jour"
          }
        },
        "required": ["username", "email"]
      },
      "Game": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri" },
          "releaseDate": { "type": "string", "format": "date" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "publisher": { "type": "string" },
          "active": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["name"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message d'erreur"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token (optionnel pour certaines routes)"
      }
    }
  },
  "paths": {
...
    "/api/users/": {
      "get": {
        "summary": "Lister tous les utilisateurs",
        "description": "Retourne la liste complète des utilisateurs (sans mots de passe)",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/games": {
      "post": {
        "summary": "Créer un jeu",
        "description": "Crée un nouveau jeu; accepte `multipart/form-data` pour upload d'image (champ `image`) ou `application/json` avec `imageUrl`.",
        "tags": ["Games"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "releaseDate": { "type": "string", "format": "date" },
                  "tags": { "oneOf": [{ "type": "string", "description": "Comma-separated tags (ex: 'action,arcade')" }, { "type": "array", "items": { "type": "string" }, "description": "Multiple tags via repeated fields (use `tags[]`)" }] },
                  "publisher": { "type": "string" }
                },
                "required": ["name"]
              }
            },
            "application/json": { "schema": { "$ref": "#/components/schemas/Game" } }
          }
        },
        "responses": {
          "201": { "description": "Jeu créé", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Game" } } } },
          "400": { "description": "Requête invalide", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Erreur serveur", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "summary": "Lister les jeux",
        "description": "Retourne la liste des jeux (pagination, filtre par tag, recherche)",
        "tags": ["Games"],
        "parameters": [
          { "name": "tag", "in": "query", "schema": { "type": "string" }, "description": "Filtrer par tag" },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Recherche par nom/description" }
        ],
        "responses": {
          "200": { "description": "Liste des jeux", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Game" } } } } },
          "500": { "description": "Erreur serveur", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}